--- src/service/user/logout.ts
@@ -1,5 +1,0 @@
-import { request } from '@/lib/clientApi/request';
-
-export const logout = async () => {
-  return await request('POST', '/api/user/logout');
-};
